+++ project.flutter/client_app/lib/page/library_page.dart
@@ -6,6 +6,7 @@ import 'package:lowresrmx/data/sync_manager.dart';
 import 'package:lowresrmx/page/manual_page.dart';
 import 'package:lowresrmx/page/settings_page.dart';
 import 'package:lowresrmx/widget/google_account_tile.dart';
+import 'package:lowresrmx/widget/sync_tile.dart';
 import 'package:lowresrmx/widget/library_grid.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:provider/provider.dart';
@@ -116,6 +117,9 @@ class MyLibraryPage extends StatelessWidget {
             builder: (context, sync, child) {
               if (sync.accountEmail == null) return const SizedBox.shrink();
               return Icon(
+								sync.syncing
+										? Icons.cloud_sync_rounded
+										:
                 sync.isAuthorized
                     ? Icons.cloud_done_rounded
                     : Icons.cloud_off_rounded,
@@ -151,6 +155,7 @@ class MyLibraryPage extends StatelessWidget {
           buildReinstallTile(context),
 					const Divider(),
           const MyGoogleAccountTile(),
+					const MySyncTile(),
           const Divider(),
           buildAboutTile(context),
         ],